Add getters for most of the members of the CVC_CERT structure.
These getters are mainly implemented in order to be able to provide nicer
bindings for higher level languages. They try to hide the underlying C
structure and the data types contained in it.

I also started renameing functions in order to provide a more consistent
interface. All public functions should be converted to begin with the prefix
"CVC_".